transaction/requirements.txt
kafka-python==2.0.2
kafka-client-decorator>=1.4